Use pytest on multi-threaded tests
The tests requiring multiple threads (one for publishing the interfaces and
one for using them) are largely rewritten to work with pytest. It doesn't
support assertions which were not asserted in the main thread. So the threads
need to either report the failed assertion or the actual result and that is
asserted in the test itself.

It also removes any modification of global variables outside the global scope,
because that would make it not possible to run multiple tests. Instead there
is a new Thread class which automatically quits the main loop after all tests
are finished. If the main loop is quit, the main thread can get the status
from the threads to check what the result was. If the thread should hang for
any reason, it'll timeout after 100 ms and report it as a failure.

Also add more tests for the class using multiple interfaces, by using only a
specific interface.

The test script itself is rewritten to call pytest only once.